Add api get list fo service package

+++ src/main/java/dts/com/vn/ilink/repository/BstLookupTableRowRepository.java
@@ -20,7 +20,7 @@ public interface BstLookupTableRowRepository extends JpaRepository<BstLookupTabl
 	@Query(nativeQuery = true, value = "SELECT max(row_id) " +
 			"FROM bst_lookup_table_row AS r " +
 			"WHERE r.table_id = ?1")
-	Long getMaxRowId(Integer tableId);
+	Long getMaxRowId(Long tableId);
 
 	@Query(nativeQuery = true, value = "SELECT r.* " +
 			"FROM bst_lookup_table_row AS r " +
